refactor: PR 코멘트 반영 #52
- Array extension(안전하게 Array Element 사용)
- DetailViewCoin 분리
