+++ BithumbYagomAcamedy/View/CoinDetailView/CoinDetailDataManager.swift
@@ -8,53 +8,11 @@
 import Foundation
 
 protocol CoinDetailDataManagerDelegate: AnyObject {
-    func coinDetailDataManager(didChange coin: CoinDetailDataManager.DetailViewCoin?)
+    func coinDetailDataManager(didChange coin: DetailViewCoin?)
 }
 
 final class CoinDetailDataManager {
-    
-    // MARK: - Nested Type
-    
-    struct DetailViewCoin {
-        let name: String
-        var price: Double?
-        var changePrice: Double?
-        var changeRate: Double?
         
-        // MARK: - DetailViewCoin Computed Property
-        
-        var commaPrice: String {
-            guard let price = price else {
-                return "오류 발생"
-            }
-            
-            let formatter = NumberFormatter()
-            formatter.numberStyle = .decimal
-            
-            return formatter.string(for: price) ?? String()
-        }
-        
-        var changePriceString: String {
-            guard let changePrice = changePrice else {
-                return "오류 발생"
-            }
-            
-            if changePrice > 0 {
-                return "+" + String(changePrice)
-            }
-            
-            return String(changePrice)
-        }
-        
-        var changeRateString: String {
-            guard let changeRate = changeRate else {
-                return "오류 발생"
-            }
-            
-            return String(changeRate) + "%"
-        }
-    }
-    
     // MARK: - Property
     
     weak var delegate: CoinDetailDataManagerDelegate?
